fix bugs in device details, refactor alert notification to not use legacy alerting
