+++ src/datasource/kentikAPI.ts
@@ -1,5 +1,3 @@
-import { appEvents } from 'grafana/app/core/core';
-
 import * as _ from 'lodash';
 import angular from 'angular';
 import { getRegion } from "./regionHelper";
@@ -130,21 +128,6 @@ export class KentikAPI {
         }
       });
   }
-}
-
-export function showAlert(error) {
-  let message = '';
-  message += error.status ? `(${error.status}) ` : '';
-  message += error.statusText ? error.statusText + ': ' : '';
-  if (error.data && error.data.error) {
-    message += error.data.error;
-  } else if (error.err) {
-    message += error.err;
-  } else if (_.isString(error)) {
-    message += error;
-  }
-
-  appEvents.emit('alert-error', ["Can't connect to Kentik API", message]);
 }
 
 angular.module('grafana.services').service('kentikAPISrv', KentikAPI);
