+++ src/panel/call-to-action/module.ts
@@ -1,4 +1,5 @@
-import { KentikAPI, showAlert } from '../../datasource/kentikAPI';
+import { KentikAPI } from '../../datasource/kentikAPI';
+import { showAlert } from "../../datasource/alertHelper";
 import { PanelCtrl, loadPluginCss } from 'grafana/app/plugins/sdk';
 import { getRegion } from "../../datasource/regionHelper";
